Rationale
This dashboard explores data on skincare products using three visualizations. The first is a scatter plot that compares price to ranking. The second measures the average prices of each type of product. The third takes the top 10 most frequently used ingredients and measures the frequency of each ingredient in each type of product.